+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]## [0.5.5.2b2]
+[Unreleased]## [0.5.5.2b3]
+
+## [0.5.5.2b2]- 2023-06-08
 ### Changed
 - Core01OptimizationEngine, Core02OptimizationEngine, PlotlyManager - added support for DataManager generics
 - Core01DataManager logger now uses `__name__`